Splitting code into fileS
